[FIX] starRating 타입 변환 및 reviewCount, bookmarkCount 로직 변환

+++ src/main/java/com/kusitms/hotsixServer/domain/review/dto/RequestReviewDto.java
@@ -11,7 +11,7 @@ public class RequestReviewDto {
 
     private Long userId;
     private Long placeId;
-    private float starRating;
+    private int starRating;
     private String content;
     private String[] stickers;
